feat: add prompts hook
