+++ src/components/prompts/PromptsPage.tsx
@@ -8,65 +8,25 @@
 
 import * as React from "react";
 import { SectionCard, Card } from "@/components/ui";
-import { usePersistentState, uid } from "@/lib/db";
 import { LOCALE } from "@/lib/utils";
 import PromptsHeader from "./PromptsHeader";
 import PromptsComposePanel from "./PromptsComposePanel";
 import PromptsDemos from "./PromptsDemos";
-
-type Prompt = {
-  id: string;
-  title?: string; // optional for back-compat
-  text: string;
-  createdAt: number;
-};
+import { usePrompts } from "./usePrompts";
 
 export default function PromptsPage() {
-  // Storage
-  const [prompts, setPrompts] = usePersistentState<Prompt[]>("prompts.v1", []);
+  const { prompts, query, setQuery, filtered, save } = usePrompts();
 
   // Drafts
   const [titleDraft, setTitleDraft] = React.useState("");
   const [textDraft, setTextDraft] = React.useState("");
 
-  // Search (now lives in the header)
-  const [query, setQuery] = React.useState("");
-
-  const deriveTitle = React.useCallback((p: Prompt) => {
-    if (p.title && p.title.trim()) return p.title.trim();
-    const firstLine = (p.text || "").split(/\r?\n/)[0]?.trim() || "";
-    return firstLine || "Untitled";
-  }, []);
-
-  // Derived: filtered list (compute titles once per prompt)
-  type PromptWithTitle = Prompt & { title: string };
-  const filtered: PromptWithTitle[] = React.useMemo(() => {
-    const q = query.trim().toLowerCase();
-    return prompts.reduce<PromptWithTitle[]>((acc, p) => {
-      const title = deriveTitle(p);
-      const text = p.text || "";
-      if (!q || title.toLowerCase().includes(q) || text.toLowerCase().includes(q)) {
-        acc.push({ ...p, title });
-      }
-      return acc;
-    }, []);
-  }, [prompts, query, deriveTitle]);
-
-  function save() {
-    const text = textDraft.trim();
-    const title = titleDraft.trim() || text.split(/\r?\n/)[0]?.trim() || "Untitled";
-    if (!text && !titleDraft.trim()) return; // nothing to save
-
-    const next: Prompt = {
-      id: uid("p"),
-      title,
-      text,
-      createdAt: Date.now(),
-    };
-    setPrompts((prev) => [next, ...prev]);
-    setTitleDraft("");
-    setTextDraft("");
-  }
+  const handleSave = React.useCallback(() => {
+    if (save(titleDraft, textDraft)) {
+      setTitleDraft("");
+      setTextDraft("");
+    }
+  }, [save, titleDraft, textDraft]);
 
   return (
     <SectionCard>
@@ -75,7 +35,7 @@ export default function PromptsPage() {
           count={prompts.length}
           query={query}
           onQueryChange={setQuery}
-          onSave={save}
+          onSave={handleSave}
           disabled={!titleDraft.trim() && !textDraft.trim()}
         />
       </SectionCard.Header>
