+++ src/components/prompts/index.ts
@@ -4,7 +4,6 @@ export { default as ButtonShowcase } from "./ButtonShowcase";
 export { default as ColorGallery } from "./ColorGallery";
 export { default as ComponentGallery } from "./ComponentGallery";
 export { default as DemoHeader } from "./DemoHeader";
-export * from "./DemoHeader";
 export { default as GalleryItem } from "./GalleryItem";
 export { default as GoalListDemo } from "./GoalListDemo";
 export { default as IconButtonShowcase } from "./IconButtonShowcase";
@@ -15,3 +14,4 @@ export { default as PromptsHeader } from "./PromptsHeader";
 export { default as PromptsPage } from "./PromptsPage";
 export { default as UpdatesList } from "./UpdatesList";
 export * from "./demoData";
+export * from "./usePrompts";
